Add first functional integration test for EditEndpoint
This test successfully uses the EditEndpoint API to create a new item,
and then asserts that it has the right reconciliation URL. To make this
possible, we also need to add some data to the EditEndpoint response.

The test can surely be improved later (e.g. extract some methods), but
it’s a start.

Bug: T281798