Inject Wikibase services into EditEndpoint (#11)
On REL1_35, they’re not yet available in the MediaWiki service
container, but getting them in the factory() method and injecting them
into the constructor is still better than getting them in the run()
method. One thing this lets us do is inject a fake property datatype
lookup in the test, which is probably more efficient than creating and
saving a real property.

We pass the PropertyDataTypeLookup into the MinimalItemInput
constructor, but apart from that the item input constructors still get
their services from the default WikibaseRepo instance. I assume we’ll
turn these into services of their own later.

Bug: T282244