Add WikibaseReconcileEditServicesTest
This tests that the lists of services in the wiring file and the
accessor class are identical, effectively ensuring that each service has
an accessor function, that the class contains no non-accessor functions
(other than a private constructor), and that the accessor functions are
in the “right” order (the same order as in the wiring file, where we
have /** @phpcs-require-sorted-array */).

Other tests can be added later, probably based on Wikibase’s
WikibaseRepoTest or WikibaseClientTest (or possibly even extracting a
base class for service wiring test files into a library?), but I didn’t
want to spend too much time on this for now.

Bug: T282551