Introduce EditRequestParser service
This turns the EditRequest interface into a simple class wrapping the
“reconcile” and “entity” values, and moves the logic that formerly lived
in UrlInputEditRequest into a new class, EditRequestParser. Later, we’ll
move more parsing logic from EditEndpoint to EditRequestParser.

In EditEndpointTest, we need to change testExecuteNoPropertyFound() to
pass an array as the “entity” parameter, rather than a string, so that
we don’t get a TypeError for the invalid “entity” parameter before even
reaching the error condition we want to test (no such property).

In the new EditRequest class, I put “reconcile” before “entity”, because
since #17 (T282232), the entity parsing eventually requires the parsed
reconciliation URL. Later in this chain of commits, we’ll also add
another field to EditRequest (the “other items”), which are more closely
related to the “entity” than the “reconcile”.

Bug: T282564