Move “invalid reconcile JSON” error to EditRequestParser
This moves the responsibility for detecting that kind of error from
EditEndpoint to EditRequestParser – EditRequest::reconcile() is now
guaranteed to return a non-null array. Additionally, reconcile payloads
that are valid JSON but not objects/arrays (e.g. strings) now produce
the same error (I think it can still fairly be described as “invalid
reconcile JSON”), instead of a TypeError when directly treating the
json_decode() result as an array or ?array.

In parseRequestInterface(), we process $reconcile before $entity because
eventually, down the road, the $entity part will require the $reconcile
result.

Bug: T282564