Move “invalid datatype” error to EditRequestParser
Checking that the reconciliation property has the "url" datatype is now
the responsibility of the EditRequestParser. This lets us remove the
propertyDataTypeLookup from EditEndpoint.

The test for a missing property is moved from EditEndpoint, though I’m
not sure if it’s intentional that we don’t catch this error and give it
a proper localized message.

Bug: T282564