Extract ItemReconciler service
This service wraps a ReconciliationService and SimplePutStrategy and
extracts the reconciliation URL from an input item before passing it to
the ReconciliationService and SimplePutStrategy. Its return value is an
Item and a revision ID, but the meaning is quite different, so we
introduce a new result class, ReconciledItem, rather than reusing
ReconciliationServiceItem.

Bug: T283567