Make some WikibaseRepo calls compatible with REL1_36
The Wikibase service migration was mostly backwards compatible, but some
methods were renamed from new*() to get()*. Check which of them exists
at runtime and call the right one.

The PHPCS rule that requires comments to be on a new line is disabled –
in Wikibase we don’t use that one either.