Preserve script path in wikibase-edit.config.js
Since maxlath/wikibase-sdk#78 and maxlath/wikibase-edit#70, those
libraries support specifying a script path, so we can now take that out
of the base URL instead of throwing it away.

npm also updated the package-lock.json to lockfileVersion 2; since this
version is compatible with older npm releases, I think that’s okay to
keep.