Set revision information on ReconciliationServiceItem
After EditEndpoint::persistItem() has saved an item, record that
information on the ReconciliationServiceItem instance, so that the next
time the same instance is returned by the ReconciliationService, we know
it doesn’t need to be saved again.

This happens on two levels: ReconciliationServiceItem gets a method to
update its contents, while keeping the instance intact, since the
instance is cached in the ReconciliationService; on the other hand,
ReconciledItem now wraps a ReconciliationServiceItem (rather than a
revision ID) and gets a method to update that while discarding itself
(since ReconciledItem instances are not long-lived).

An integration test is added which checks that the revision information
makes it “into” the ReconciliationService. (The test doesn’t require
that the ReconciliationService returns the same instance – I assume this
could also be fulfilled by saving a reference to the service in the
individual items and having updateFromEntityRevision() send the
EntityRevision to the service without updating itself.)

Bug: T284068