Stop tracking other items in BatchEditEndpoint
With the changes in the previous commit, we already know at all times
which items were saved and which weren’t, so we no longer need to keep
track of it here. This also means we can stop returning those items in
persistItem(), and only return a Status for the main item.

Bug: T284068